:sparkles: feat: 添加 3D 模型 Tab
