+++ src/constants/token.ts
@@ -18,5 +18,6 @@ export const LIST_GRID_HEIGHT = 108;
 export const INPUT_WIDTH_S = 240;
 export const INPUT_WIDTH_M = 320;
 export const INPUT_WIDTH_L = 480;
-
 export const INPUT_WIDTH_XL = 640;
+
+export const ROLE_VIEWER_HEIGHT = 640;
