+++ src/features/AgentViewer/Role/style.ts
@@ -1,9 +1,11 @@
 import { createStyles } from 'antd-style';
 
+import { ROLE_VIEWER_HEIGHT } from '@/constants/token';
+
 export const useStyles = createStyles(({ css, token }) => ({
   viewer: css`
     position: relative;
-    height: 100%;
+    height: ${ROLE_VIEWER_HEIGHT}px;
     min-height: 0;
   `,
   toolbar: css`
@@ -19,8 +21,10 @@ export const useStyles = createStyles(({ css, token }) => ({
   `,
   canvas: css`
     display: block;
+
     width: 100%;
     max-width: 100%;
+    height: 100%;
     max-height: 100%;
   `,
 }));
