--- src/panels/RolePanel/Touch/SideBar/index.tsx
@@ -1,7 +1,0 @@
-import AreaList from './AreaList';
-
-const SideBar = () => {
-  return <AreaList />;
-};
-
-export default SideBar;
